+++ atlasdb-config/src/main/java/com/palantir/atlasdb/factory/timelock/clock/ClockSkewMonitor.java
@@ -1,5 +1,5 @@
 /*
- * (c) Copyright 2018 Palantir Technologies Inc. All rights reserved.
+ * (c) Copyright 2019 Palantir Technologies Inc. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.palantir.atlasdb.timelock.clock;
+package com.palantir.atlasdb.factory.timelock.clock;
 
 import java.time.Duration;
 import java.time.temporal.ChronoUnit;
